Remove custom keylayout.
I was reluctant to add this in the first place. It's really brittle
and gross. The nail in the coffin is that the custom input source
does not persist after restarts, requiring me to select the "mtse"
input source for it to actually work. I will just work on making new
VS code bindings for meta n and meta p.